refactor: split timeout-salvage helpers out of agent_service_recovery
agent_service_recovery.py reached 1562 lines and tripped the same
first-party line-limit guardrail as the probe test module. Move the
self-contained timeout-salvage block — _masked_agent_timeout_reason_code,
_record_timeout_rerun_floor, _secure_timeout_rerun_preservation and
_timeout_rerun_salvage_steps — into a focused sibling module, following
the package's existing one-concern-per-module layout.

The two names the recovery loop calls are imported back, so every existing
importer and the RunnerDelegatesMixin wiring are untouched. The moved code
is byte-identical and only the imports that travelled with it (asyncio,
worktree_writer_locks_borrowed_from) leave the original module; no
behaviour change.
